23 01 18 인성 버그 수정

+++ FOODLE/src/main/java/com/smhrd/model/DailyWeightVO.java
@@ -19,7 +19,7 @@ public class DailyWeightVO {
 	
 	
 	@NonNull
-	private int curr_weight;
+	private Float curr_weight;
 	private String mb_id;
 	private String date;
